requirements.txt
asgiref==3.3.4
astroid==2.5.5
Brotli==1.0.9
colorama==0.4.4
dj-database-url==0.5.0
dj-email-url==1.0.2
Django==3.2
django-cache-url==3.2.3
django-ckeditor==6.0.0
django-heroku==0.3.1
django-js-asset==1.2.2
environs==9.3.2
gunicorn==20.1.0
isort==5.8.0
lazy-object-proxy==1.6.0
marshmallow==3.11.1
mccabe==0.6.1
Pillow==8.2.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
pylint==2.8.0
python-dotenv==0.17.1
pytz==2021.1
sqlparse==0.4.1
toml==0.10.2
whitenoise==5.2.0
wrapt==1.12.1
